requirements.txt
requests==2.31.0
pandas==2.1.4
nltk==3.8.1
beautifulsoup4==4.12.2
biopython
click
